fix post merge errors and tests
