+++ packages/lodestar/src/network/peers/interface.ts
@@ -7,6 +7,12 @@ export interface IPeerMetadataStore {
   getStatus(peer: PeerId): Status | null;
   setScore(peer: PeerId, score: number | null): void;
   getScore(peer: PeerId): number | null;
+
+  /**
+   * It will never store metadata with lower seq number
+   * @param peer
+   * @param metadata
+   */
   setMetadata(peer: PeerId, metadata: Metadata | null): void;
   getMetadata(peer: PeerId): Metadata | null;
   setEncoding(peer: PeerId, encoding: ReqRespEncoding | null): void;
