+++ packages/lodestar/src/sync/reqResp/reqResp.ts
@@ -24,7 +24,7 @@ import {computeStartSlotAtEpoch, GENESIS_SLOT, getBlockRootAtSlot} from "@chains
 import {toHexString} from "@chainsafe/ssz";
 import {RpcError} from "../../network/error";
 import {createStatus, syncPeersStatus} from "../utils/sync";
-import {handlePeerMetadataSequence} from "../utils/reputation";
+import {handlePeerMetadataSequence} from "../../network/peers/utils";
 
 export interface IReqRespHandlerModules {
   config: IBeaconConfig;
@@ -51,7 +51,7 @@ export class BeaconReqRespHandler implements IReqRespHandler {
   private network: INetwork;
   private logger: ILogger;
 
-  public constructor({config, db, chain, network, reputationStore, logger}: IReqRespHandlerModules) {
+  public constructor({config, db, chain, network, logger}: IReqRespHandlerModules) {
     this.config = config;
     this.db = db;
     this.chain = chain;
@@ -207,7 +207,7 @@ export class BeaconReqRespHandler implements IReqRespHandler {
   public async onPing(peerId: PeerId, id: RequestId, request: Ping): Promise<void> {
     this.network.reqResp.sendResponse(id, null, this.network.metadata.seqNumber);
     // no need to wait
-    handlePeerMetadataSequence(this.reps, this.network.reqResp, this.logger, peerId, request).catch(() => {
+    handlePeerMetadataSequence(this.network, this.logger, peerId, request).catch(() => {
       this.logger.warn(`Failed to handle peer ${peerId.toB58String()} metadata sequence ${request}`);
     });
   }
@@ -272,7 +272,6 @@ export class BeaconReqRespHandler implements IReqRespHandler {
       const request = createStatus(this.chain);
       try {
         this.network.peerMetadata.setStatus(peerId, await this.network.reqResp.status(peerId, request));
-        this.reps.get(peerId.toB58String()).latestMetadata = await this.network.reqResp.metadata(peerId);
       } catch (e) {
         this.logger.verbose(`Failed to get peer ${peerId.toB58String()} latest status and metadata`, {
           reason: e.message,
