+++ packages/lodestar/test/unit/network/tasks/checkPeerAliveTask.test.ts
@@ -1,28 +1,29 @@
 import sinon, {SinonStubbedInstance} from "sinon";
 import {CheckPeerAliveTask} from "../../../../src/network/tasks/checkPeerAliveTask";
 import {INetwork, IReqResp, Libp2pNetwork} from "../../../../src/network";
-import {ReputationStore} from "../../../../src/sync/IReputation";
 import {ReqResp} from "../../../../src/network/reqResp";
 import {WinstonLogger} from "@chainsafe/lodestar-utils";
 import {config} from "@chainsafe/lodestar-config/lib/presets/mainnet";
 import {expect} from "chai";
 import PeerId from "peer-id";
+import {IPeerMetadataStore} from "../../../../src/network/peers/interface";
+import {Libp2pPeerMetadataStore} from "../../../../src/network/peers/metastore";
 
 describe("CheckPeerAliveTask", function () {
   let networkStub: SinonStubbedInstance<INetwork>;
   let reqRespStub: SinonStubbedInstance<IReqResp>;
-  let reps: ReputationStore;
+  let peerMetadataStub: SinonStubbedInstance<IPeerMetadataStore>;
   let task: CheckPeerAliveTask;
   let peerId: PeerId;
   beforeEach(async () => {
     networkStub = sinon.createStubInstance(Libp2pNetwork);
     reqRespStub = sinon.createStubInstance(ReqResp);
     networkStub.reqResp = reqRespStub;
-    reps = new ReputationStore();
+    peerMetadataStub = sinon.createStubInstance(Libp2pPeerMetadataStore);
+    networkStub.peerMetadata = peerMetadataStub;
     task = new CheckPeerAliveTask(config, {
       logger: new WinstonLogger(),
       network: networkStub,
-      reps,
     });
     // @ts-ignore
     networkStub.metadata = {
@@ -52,10 +53,10 @@ describe("CheckPeerAliveTask", function () {
 
   it("ping successfully, return same sequence number", async () => {
     reqRespStub.ping.resolves(BigInt(1));
-    reps.getFromPeerId(peerId).latestMetadata = {
+    peerMetadataStub.getMetadata.withArgs(peerId).returns({
       seqNumber: BigInt(1),
       attnets: Array(64).fill(true),
-    };
+    });
     await task.run();
     expect(networkStub.disconnect.called).to.be.false;
     expect(reqRespStub.metadata.called).to.be.false;
@@ -63,10 +64,10 @@ describe("CheckPeerAliveTask", function () {
 
   it("ping successfully, return bigger sequence number", async () => {
     reqRespStub.ping.resolves(BigInt(10));
-    reps.getFromPeerId(peerId).latestMetadata = {
+    peerMetadataStub.getMetadata.withArgs(peerId).returns({
       seqNumber: BigInt(1),
       attnets: Array(64).fill(true),
-    };
+    });
     await task.run();
     expect(networkStub.disconnect.called).to.be.false;
     expect(reqRespStub.metadata.calledOnce).to.be.true;
